requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.5
langchain-community>=0.0.10
python-dotenv>=1.0.0
streamlit-authenticator>=0.2.3
bcrypt>=4.0.0
PyJWT>=2.8.0
requests>=2.31.0
tomli